backend/pyproject.toml
[project]
name = "mmm-backend"
version = "0.1.0"
description = "Marketing Mix Modeling Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "greenlet>=3.0.0",
    "alembic>=1.13.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
    "statsmodels>=0.14.0",
    "resend>=0.7.0",
    "reportlab>=4.0.0",
    "python-pptx>=0.6.21",
    "gunicorn>=23.0.0",
    "pymc>=5.10.0",
    "arviz>=0.17.0",
    "sentry-sdk[fastapi,celery]>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 120
extend-exclude = ["alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = [
    "N806",  # Variable in function should be lowercase (ML convention: X, Y)
    "N803",  # Argument name should be lowercase (ML convention)
    "E741",  # Ambiguous variable name (ML convention: l, I)
    "F821",  # Undefined name (SQLAlchemy string references)
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
